+++ information_gain_aml/__init__.py
@@ -1,3 +1,3 @@
 """Information Gain-based Online Action Model Learning."""
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"